.github: add package only PAT token and update ops.yaml
> Note: By default, when you select the write:packages scope for your personal
access token (PAT) in the user interface, the repo scope will also be selected.
The repo scope offers unnecessary and broad access, which we recommend you
avoid using for GitHub Actions workflows in particular. For more information,
see "Security hardening for GitHub Actions." As a workaround, you can select
just the write:packages scope for your PAT in the user interface with this url:
https://github.com/settings/tokens/new?scopes=write:packages.

References: https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-container-registry#authenticating-to-the-container-registry

Signed-off-by: Wei Fu <fuweid89@gmail.com>